+++ src/signal.c
@@ -118,7 +118,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        GPS_CA_CHIPPING_RATE,
                        false,
                        GPS_L2CM_PRN_PERIOD_MS,
-                       GPS_L2_DOPPLER_MAX_HZ,
+                       (float)GPS_L2_DOPPLER_MAX_HZ,
                        -0.25f, /* see L2S in [1] */
                        true},
     [CODE_GPS_L2CL] = {CONSTELLATION_GPS,
@@ -131,7 +131,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        GPS_CA_CHIPPING_RATE,
                        false,
                        GPS_L2CL_PRN_PERIOD_MS,
-                       GPS_L2_DOPPLER_MAX_HZ,
+                       (float)GPS_L2_DOPPLER_MAX_HZ,
                        -0.25f, /* see L2L in [1] */
                        false},
     [CODE_GPS_L2CX] = {CONSTELLATION_GPS,
@@ -144,7 +144,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        0,
                        false,
                        0,
-                       GPS_L2_DOPPLER_MAX_HZ,
+                       (float)GPS_L2_DOPPLER_MAX_HZ,
                        -0.25f, /* see L2X [1] */
                        false},
     [CODE_GPS_L5I] = {CONSTELLATION_GPS,
@@ -157,7 +157,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       GPS_L5_CHIPPING_RATE,
                       false,
                       GPS_L5_PRN_PERIOD_MS,
-                      GPS_L5_DOPPLER_MAX_HZ,
+                      (float)GPS_L5_DOPPLER_MAX_HZ,
                       0.f, /* reference signal (see L5I in [1]) */
                       false},
     [CODE_GPS_L5Q] = {CONSTELLATION_GPS,
@@ -170,7 +170,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       0,
                       false,
                       0,
-                      GPS_L5_DOPPLER_MAX_HZ,
+                      (float)GPS_L5_DOPPLER_MAX_HZ,
                       -0.25f, /* see L5Q in [1] */
                       false},
     [CODE_GPS_L5X] = {CONSTELLATION_GPS,
@@ -183,7 +183,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       0,
                       false,
                       0,
-                      GPS_L5_DOPPLER_MAX_HZ,
+                      (float)GPS_L5_DOPPLER_MAX_HZ,
                       0.f, /* must be aligned to CODE_GPS_L5I (see L5X in [1])*/
                       false},
     [CODE_GPS_L1P] = {CONSTELLATION_GPS,
@@ -209,7 +209,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       0,
                       false,
                       0,
-                      GPS_L2_DOPPLER_MAX_HZ,
+                      (float)GPS_L2_DOPPLER_MAX_HZ,
                       0.f, /* reference signal (see L2P in [1]) */
                       false},
 
@@ -250,7 +250,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        SBAS_L5_CHIPPING_RATE,
                        false,
                        SBAS_L5_PRN_PERIOD_MS,
-                       SBAS_L5_DOPPLER_MAX_HZ,
+                       (float)SBAS_L5_DOPPLER_MAX_HZ,
                        0.f, /* reference signal */
                        true},
     [CODE_SBAS_L5Q] = {CONSTELLATION_SBAS,
@@ -263,7 +263,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        SBAS_L5_CHIPPING_RATE,
                        false,
                        SBAS_L5_PRN_PERIOD_MS,
-                       SBAS_L5_DOPPLER_MAX_HZ,
+                       (float)SBAS_L5_DOPPLER_MAX_HZ,
                        -0.25f, /* not used */
                        false},
     [CODE_SBAS_L5X] = {CONSTELLATION_SBAS,
@@ -276,7 +276,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        SBAS_L5_CHIPPING_RATE,
                        false,
                        SBAS_L5_PRN_PERIOD_MS,
-                       SBAS_L5_DOPPLER_MAX_HZ,
+                       (float)SBAS_L5_DOPPLER_MAX_HZ,
                        0.f, /* must be aligned with CODE_SBAS_L5I */
                        false},
 
@@ -291,7 +291,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        GLO_CA_CHIPPING_RATE,
                        true,
                        GLO_PRN_PERIOD_MS,
-                       GLO_L1_DOPPLER_MAX_HZ,
+                       (float)GLO_L1_DOPPLER_MAX_HZ,
                        0.f, /* reference signal (see L1C in [1]) */
                        true},
     [CODE_GLO_L2OF] = {CONSTELLATION_GLO,
@@ -304,7 +304,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        GLO_CA_CHIPPING_RATE,
                        false,
                        GLO_PRN_PERIOD_MS,
-                       GLO_L2_DOPPLER_MAX_HZ,
+                       (float)GLO_L2_DOPPLER_MAX_HZ,
                        0.f, /* reference signal (see L2C in [1]) */
                        true},
     [CODE_GLO_L1P] = {CONSTELLATION_GLO,
@@ -317,7 +317,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       0,
                       false,
                       0,
-                      GLO_L1_DOPPLER_MAX_HZ,
+                      (float)GLO_L1_DOPPLER_MAX_HZ,
                       0.25f, /* see L1P in [1] */
                       false},
     [CODE_GLO_L2P] = {CONSTELLATION_GLO,
@@ -330,7 +330,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       0,
                       false,
                       0,
-                      GLO_L2_DOPPLER_MAX_HZ,
+                      (float)GLO_L2_DOPPLER_MAX_HZ,
                       0.25f, /* see L2P in [1] */
                       false},
 
@@ -397,7 +397,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       GAL_E6_CHIPPING_RATE,
                       false,
                       GAL_E6B_PRN_PERIOD_MS,
-                      GAL_E6_DOPPLER_MAX_HZ,
+                      (float)GAL_E6_DOPPLER_MAX_HZ,
                       0.f, /* reference signal (see L6B in [1]) */
                       true},
     [CODE_GAL_E6C] = {CONSTELLATION_GAL,
@@ -410,7 +410,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       0,
                       false,
                       0,
-                      GAL_E6_DOPPLER_MAX_HZ,
+                      (float)GAL_E6_DOPPLER_MAX_HZ,
                       -0.5f, /* see L6C in [1] */
                       false},
     [CODE_GAL_E6X] = {CONSTELLATION_GAL,
@@ -423,7 +423,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       0,
                       false,
                       0,
-                      GAL_E6_DOPPLER_MAX_HZ,
+                      (float)GAL_E6_DOPPLER_MAX_HZ,
                       0.f, /* must be aligned to CODE_GAL_E6B (see L6X in [1])*/
                       false},
     [CODE_GAL_E7I] = {CONSTELLATION_GAL,
@@ -436,7 +436,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       GAL_E7_CHIPPING_RATE,
                       false,
                       GAL_E7I_PRN_PERIOD_MS,
-                      GAL_E7_DOPPLER_MAX_HZ,
+                      (float)GAL_E7_DOPPLER_MAX_HZ,
                       0.f, /* reference signal (see L7I in [1]) */
                       true},
     [CODE_GAL_E7Q] = {CONSTELLATION_GAL,
@@ -449,7 +449,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       0,
                       false,
                       0,
-                      GAL_E7_DOPPLER_MAX_HZ,
+                      (float)GAL_E7_DOPPLER_MAX_HZ,
                       -0.25f, /* see L7Q in [1] */
                       false},
     [CODE_GAL_E7X] = {CONSTELLATION_GAL,
@@ -462,7 +462,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       0,
                       false,
                       0,
-                      GAL_E7_DOPPLER_MAX_HZ,
+                      (float)GAL_E7_DOPPLER_MAX_HZ,
                       0.f, /* must be aligned to CODE_GAL_E7I (see L7X in [1])*/
                       false},
     [CODE_GAL_E8I] = {CONSTELLATION_GAL,
@@ -475,7 +475,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       0,
                       false,
                       0,
-                      GAL_E8_DOPPLER_MAX_HZ,
+                      (float)GAL_E8_DOPPLER_MAX_HZ,
                       0.f, /* reference signal (see L8I in [1]) */
                       false},
     [CODE_GAL_E8Q] = {CONSTELLATION_GAL,
@@ -488,7 +488,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       0,
                       false,
                       0,
-                      GAL_E8_DOPPLER_MAX_HZ,
+                      (float)GAL_E8_DOPPLER_MAX_HZ,
                       -0.25f, /* see L8Q in [1] */
                       false},
     [CODE_GAL_E8X] = {CONSTELLATION_GAL,
@@ -501,7 +501,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       0,
                       false,
                       0,
-                      GAL_E8_DOPPLER_MAX_HZ,
+                      (float)GAL_E8_DOPPLER_MAX_HZ,
                       0.f, /* must be aligned to CODE_GAL_E8Q (see L8X in [1])*/
                       false},
     [CODE_GAL_E5I] = {CONSTELLATION_GAL,
@@ -514,7 +514,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       GAL_E5_CHIPPING_RATE,
                       false,
                       GAL_E5I_PRN_PERIOD_MS,
-                      GAL_E5_DOPPLER_MAX_HZ,
+                      (float)GAL_E5_DOPPLER_MAX_HZ,
                       0.f, /* reference signal (see L5I in [1]) */
                       true},
     [CODE_GAL_E5Q] = {CONSTELLATION_GAL,
@@ -527,7 +527,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       0,
                       false,
                       0,
-                      GAL_E5_DOPPLER_MAX_HZ,
+                      (float)GAL_E5_DOPPLER_MAX_HZ,
                       -0.25f, /* see L5Q in [1] */
                       false},
     [CODE_GAL_E5X] = {CONSTELLATION_GAL,
@@ -540,7 +540,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       0,
                       false,
                       0,
-                      GAL_E5_DOPPLER_MAX_HZ,
+                      (float)GAL_E5_DOPPLER_MAX_HZ,
                       0.f, /* must be aligned to CODE_GAL_E5I (see L5X in [1])*/
                       false},
 
@@ -581,7 +581,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       BDS2_B2_CHIPPING_RATE,
                       false,
                       BDS2_B2_PRN_PERIOD_MS,
-                      BDS2_B2_DOPPLER_MAX_HZ,
+                      (float)BDS2_B2_DOPPLER_MAX_HZ,
                       0.f, /* reference signal (see L7I in [1]) */
                       true},
     [CODE_BDS3_B1CI] = {CONSTELLATION_BDS,
@@ -594,7 +594,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                         BDS3_B1C_CHIPPING_RATE,
                         false,
                         BDS3_B1C_PRN_PERIOD_MS,
-                        BDS3_B1C_DOPPLER_MAX_HZ,
+                        (float)BDS3_B1C_DOPPLER_MAX_HZ,
                         0.f, /* not used (interoperable with SBAS) */
                         false},
     [CODE_BDS3_B1CQ] = {CONSTELLATION_BDS,
@@ -607,7 +607,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                         0,
                         false,
                         0,
-                        BDS3_B1C_DOPPLER_MAX_HZ,
+                        (float)BDS3_B1C_DOPPLER_MAX_HZ,
                         0.f, /* not used (interoperable with SBAS) */
                         false},
     [CODE_BDS3_B1CX] = {CONSTELLATION_BDS,
@@ -620,7 +620,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                         0,
                         false,
                         0,
-                        BDS3_B1C_DOPPLER_MAX_HZ,
+                        (float)BDS3_B1C_DOPPLER_MAX_HZ,
                         0.f, /* not used (interoperable with SBAS) */
                         false},
     [CODE_BDS3_B3I] = {CONSTELLATION_BDS,
@@ -633,7 +633,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        BDS3_B3_CHIPPING_RATE,
                        false,
                        BDS3_B3_PRN_PERIOD_MS,
-                       BDS3_B3_DOPPLER_MAX_HZ,
+                       (float)BDS3_B3_DOPPLER_MAX_HZ,
                        0.f, /* reference signal (see L6I in [1]) */
                        false},
     [CODE_BDS3_B3Q] = {CONSTELLATION_BDS,
@@ -646,7 +646,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        0,
                        false,
                        0,
-                       BDS3_B3_DOPPLER_MAX_HZ,
+                       (float)BDS3_B3_DOPPLER_MAX_HZ,
                        -0.25f, /* see L6Q in [1] */
                        false},
     [CODE_BDS3_B3X] = {CONSTELLATION_BDS,
@@ -659,7 +659,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        0,
                        false,
                        0,
-                       BDS3_B3_DOPPLER_MAX_HZ,
+                       (float)BDS3_B3_DOPPLER_MAX_HZ,
                        0.f, /*must be aligned to CODE_BDS3_B3I (L6X in [1]) */
                        false},
     [CODE_BDS3_B7I] = {CONSTELLATION_BDS,
@@ -672,7 +672,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        BDS3_B7_CHIPPING_RATE,
                        false,
                        BDS3_B7_PRN_PERIOD_MS,
-                       BDS3_B7_DOPPLER_MAX_HZ,
+                       (float)BDS3_B7_DOPPLER_MAX_HZ,
                        0.f, /* reference signal (see L7I in [1]) */
                        false},
     [CODE_BDS3_B7Q] = {CONSTELLATION_BDS,
@@ -685,7 +685,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        0,
                        false,
                        0,
-                       BDS3_B7_DOPPLER_MAX_HZ,
+                       (float)BDS3_B7_DOPPLER_MAX_HZ,
                        -0.25f, /* see L7Q in [1] */
                        false},
     [CODE_BDS3_B7X] = {CONSTELLATION_BDS,
@@ -698,7 +698,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        BDS3_B7_CHIPPING_RATE,
                        false,
                        BDS3_B7_PRN_PERIOD_MS,
-                       BDS3_B7_DOPPLER_MAX_HZ,
+                       (float)BDS3_B7_DOPPLER_MAX_HZ,
                        0.f, /* must be aligned to CODE_BDS3_B7I (L7X in [1]) */
                        false},
     [CODE_BDS3_B5I] = {CONSTELLATION_BDS,
@@ -711,7 +711,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        BDS3_B5_CHIPPING_RATE,
                        false,
                        BDS3_B5_PRN_PERIOD_MS,
-                       BDS3_B5_DOPPLER_MAX_HZ,
+                       (float)BDS3_B5_DOPPLER_MAX_HZ,
                        0.f, /* not used (interoperable with SBAS) */
                        false},
     [CODE_BDS3_B5Q] = {CONSTELLATION_BDS,
@@ -724,7 +724,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        0,
                        false,
                        0,
-                       BDS3_B5_DOPPLER_MAX_HZ,
+                       (float)BDS3_B5_DOPPLER_MAX_HZ,
                        0.f, /* not used (interoperable with SBAS) */
                        false},
     [CODE_BDS3_B5X] = {CONSTELLATION_BDS,
@@ -737,7 +737,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        0,
                        false,
                        0,
-                       BDS3_B5_DOPPLER_MAX_HZ,
+                       (float)BDS3_B5_DOPPLER_MAX_HZ,
                        0.f, /* not used (interoperable with SBAS) */
                        false},
 
@@ -778,7 +778,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        GPS_CA_CHIPPING_RATE,
                        false,
                        GPS_L1C_PRN_PERIOD_MS,
-                       GPS_L1_DOPPLER_MAX_HZ,
+                       (float)GPS_L1_DOPPLER_MAX_HZ,
                        0.f, /* see L1S in [1] */
                        false},
     [CODE_QZS_L1CQ] = {CONSTELLATION_QZS,
@@ -791,7 +791,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        0,
                        false,
                        0,
-                       GPS_L1_DOPPLER_MAX_HZ,
+                       (float)GPS_L1_DOPPLER_MAX_HZ,
                        0.25f, /* see L1L in [1] */
                        false},
     [CODE_QZS_L1CX] = {CONSTELLATION_QZS,
@@ -804,7 +804,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        0,
                        false,
                        0,
-                       GPS_L1_DOPPLER_MAX_HZ,
+                       (float)GPS_L1_DOPPLER_MAX_HZ,
                        0.25f, /* see L1X in [1] */
                        false},
     [CODE_QZS_L2CM] = {CONSTELLATION_QZS,
@@ -817,7 +817,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        QZS_L1CA_CHIPPING_RATE,
                        false,
                        GPS_L2CM_PRN_PERIOD_MS,
-                       QZS_L2_DOPPLER_MAX_HZ,
+                       (float)QZS_L2_DOPPLER_MAX_HZ,
                        0.f, /* reference signal (see L2S in [1]) */
                        true},
     [CODE_QZS_L2CL] = {CONSTELLATION_QZS,
@@ -830,7 +830,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        QZS_L1CA_CHIPPING_RATE,
                        false,
                        GPS_L2CL_PRN_PERIOD_MS,
-                       QZS_L2_DOPPLER_MAX_HZ,
+                       (float)QZS_L2_DOPPLER_MAX_HZ,
                        0.f, /* see L2L in [1] */
                        false},
     [CODE_QZS_L2CX] = {CONSTELLATION_QZS,
@@ -843,7 +843,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                        0,
                        false,
                        0,
-                       QZS_L2_DOPPLER_MAX_HZ,
+                       (float)QZS_L2_DOPPLER_MAX_HZ,
                        0.f, /* see L2X in [1] */
                        false},
     [CODE_QZS_L5I] = {CONSTELLATION_QZS,
@@ -856,7 +856,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       0,
                       false,
                       0,
-                      QZS_L5_DOPPLER_MAX_HZ,
+                      (float)QZS_L5_DOPPLER_MAX_HZ,
                       0.f, /* reference signal (see L5I in [1]) */
                       false},
     [CODE_QZS_L5Q] = {CONSTELLATION_QZS,
@@ -869,7 +869,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       0,
                       false,
                       0,
-                      QZS_L5_DOPPLER_MAX_HZ,
+                      (float)QZS_L5_DOPPLER_MAX_HZ,
                       -0.25f, /* see L5Q in [1] */
                       false},
     [CODE_QZS_L5X] = {CONSTELLATION_QZS,
@@ -882,7 +882,7 @@ static const code_table_element_t code_table[CODE_COUNT] = {
                       0,
                       false,
                       0,
-                      QZS_L5_DOPPLER_MAX_HZ,
+                      (float)QZS_L5_DOPPLER_MAX_HZ,
                       0.f, /* must be aligned to CODE_QZS_L5I (see L5X in [1])*/
                       false},
 };
